% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_kagi_query.R
\name{build_kagi_query}
\alias{build_kagi_query}
\title{Build a Kagi search query string}
\usage{
build_kagi_query(
  query,
  filetype = NULL,
  site = NULL,
  inurl = NULL,
  intitle = NULL,
  expand = TRUE,
  open_in_browser = FALSE
)
}
\arguments{
\item{query}{Character scalar with the main search terms (can include quotes/\code{AND}/\code{OR}/...).}

\item{filetype}{Character vector of file extensions, e.g. c("pdf","docx"). Combined with \code{OR.}}

\item{site}{Character vector of domains, e.g. c("example.com","gov"). Combined with \code{OR.}}

\item{inurl}{Character vector of URL substrings to require in the URL. Multiple act like \code{AND.}}

\item{intitle}{Character vector of title terms to require in the page title. Multiple act like \code{AND.}}

\item{expand}{\code{TRUE} or \code{FALSE}, whether to expand the query to include all combinations of the parameters.
Default: \code{TRUE}. If \code{FALSE}, the normal R recycling rules are used - so use with care!}

\item{open_in_browser}{\code{TRUE} or \code{FALSE}, whether to open the query in the browser. Default: \code{FALSE}}
}
\value{
A single character string suitable for the \code{q} parameter.
}
\description{
Build a Kagi search query string
}
