% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_simple.R
\name{kagi_enrich_once}
\alias{kagi_enrich_once}
\title{Perform a Kagi enrich in one call}
\usage{
kagi_enrich_once(
  q,
  api_key = Sys.getenv("KAGI_API_KEY"),
  base_url = "https://kagi.com/api/v0"
)
}
\arguments{
\item{q}{enrich query}

\item{api_key}{Kagi API key}

\item{base_url}{Base URL for the Kagi API}

\item{limit}{Maximum number of results}
}
\value{
A kagi_enrich_results object
}
\description{
Perform a Kagi enrich in one call
}
