% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_perform.R, R/search_perform.R
\name{kagi_perform}
\alias{kagi_perform}
\title{Execute a Kagi API request}
\usage{
kagi_perform(x, ...)

kagi_perform(x, ...)
}
\arguments{
\item{x}{The request object to execute}

\item{...}{Additional parameters passed to methods}

\item{path}{Optional file path to save the raw JSON response}
}
\value{
A results object corresponding to the request type

A results object corresponding to the request type
}
\description{
Execute a Kagi API request

Execute a Kagi API request
}
