% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_simple.R
\name{kagi_search_once}
\alias{kagi_search_once}
\title{Perform a Kagi search in one call}
\usage{
kagi_search_once(
  q,
  limit = NULL,
  api_key = Sys.getenv("KAGI_API_KEY"),
  base_url = "https://kagi.com/api/v0"
)
}
\arguments{
\item{q}{Search query}

\item{limit}{Maximum number of results}

\item{api_key}{Kagi API key}

\item{base_url}{Base URL for the Kagi API}
}
\value{
A kagi_search_results object
}
\description{
Perform a Kagi search in one call
}
