% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{new_kagi_connection}
\alias{new_kagi_connection}
\title{Create a new Kagi API connection}
\usage{
new_kagi_connection(
  base_url = "https://kagi.com/api/v0",
  endpoint = as.character(NA),
  api_key = Sys.getenv("KAGI_API_KEY"),
  max_tries = 3
)
}
\arguments{
\item{base_url}{Base URL for the Kagi API}

\item{endpoint}{API endpoint to use (e.g., "search", "summarizes" or "enrich"). Default: NA}

\item{api_key}{Kagi API key}
}
\value{
A kagi_connection object
}
\description{
Create a new Kagi API connection
}
