% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{new_kagi_enrich}
\alias{new_kagi_enrich}
\title{Create a new Kagi enrich request}
\usage{
new_kagi_enrich(conn, q, limit = NULL)
}
\arguments{
\item{conn}{A kagi_connection object}

\item{q}{enrich query}

\item{limit}{Maximum number of results to return}
}
\value{
A kagi_enrich object
}
\description{
Create a new Kagi enrich request
}
