% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_results.R
\name{new_kagi_enrich_results}
\alias{new_kagi_enrich_results}
\title{Create a new Kagi enrich results object}
\usage{
new_kagi_enrich_results(enrich, raw_json, parsed)
}
\arguments{
\item{enrich}{A kagi_enrich object}

\item{raw}{The raw JSON response from the API}
}
\value{
A kagi_enrich_results object
}
\description{
Create a new Kagi enrich results object
}
