% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{new_kagi_search}
\alias{new_kagi_search}
\title{Create a new Kagi search request}
\usage{
new_kagi_search(conn, q, limit = NULL)
}
\arguments{
\item{conn}{A kagi_connection object}

\item{q}{Search query}

\item{limit}{Maximum number of results to return}
}
\value{
A kagi_search object
}
\description{
Create a new Kagi search request
}
