% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_results.R
\name{new_kagi_search_results}
\alias{new_kagi_search_results}
\title{Create a new Kagi search results object}
\usage{
new_kagi_search_results(search, raw_json, parsed)
}
\arguments{
\item{search}{A kagi_search object}

\item{raw}{The raw JSON response from the API}
}
\value{
A kagi_search_results object
}
\description{
Create a new Kagi search results object
}
